refactor: split web handler file closes #184
