+++ internal/web/handlers_entity.go
@@ -1,7 +1,6 @@
 package web
 
 import (
-	"bytes"
 	"context"
 	"errors"
 	"fmt"
@@ -11,121 +10,9 @@ import (
 	"github.com/asphaltbuffet/wherehouse/internal/app"
 	"github.com/asphaltbuffet/wherehouse/internal/inventory"
 	"github.com/asphaltbuffet/wherehouse/internal/store"
-	"github.com/asphaltbuffet/wherehouse/internal/version"
 )
 
-const (
-	searchLimit   = 50
-	maxQueryLen   = 100
-	maxFormBytes  = 16 * 1024 // upper bound for POST form bodies
-	htmxHeaderVal = "true"
-)
-
-// serverError logs the internal detail and returns a generic 500 to the client.
-// Use this for any failure that originates inside the server (DB, template, app
-// layer) — never expose the wrapped error text in the response.
-func (s *Server) serverError(w http.ResponseWriter, op string, err error) {
-	s.cfg.Logger.Error(op, "error", err)
-	http.Error(w, "internal error", http.StatusInternalServerError)
-}
-
-// renderHTML renders one template into a buffer and writes it to w with the
-// HTML content-type. On template error the response is a clean 500 — the
-// partial output never reaches the client.
-func (s *Server) renderHTML(w http.ResponseWriter, name string, data any) {
-	var buf bytes.Buffer
-	if err := s.templates.ExecuteTemplate(&buf, name, data); err != nil {
-		s.serverError(w, "execute "+name, err)
-		return
-	}
-	w.Header().Set("Content-Type", "text/html; charset=utf-8")
-	if _, err := w.Write(buf.Bytes()); err != nil {
-		s.cfg.Logger.Error("write response", "error", err, "template", name)
-	}
-}
-
-// renderHTMLAppend renders into a buffer and appends to w without setting
-// Content-Type or status. Used for HTMX out-of-band swaps that follow a
-// primary render. Template errors are logged but the primary response is
-// not affected.
-func (s *Server) renderHTMLAppend(w http.ResponseWriter, name string, data any) {
-	var buf bytes.Buffer
-	if err := s.templates.ExecuteTemplate(&buf, name, data); err != nil {
-		s.cfg.Logger.Error("execute "+name, "error", err)
-		return
-	}
-	if _, err := w.Write(buf.Bytes()); err != nil {
-		s.cfg.Logger.Error("write OOB", "error", err, "template", name)
-	}
-}
-
-// handleHealthz responds with 200 OK for liveness checks.
-func (s *Server) handleHealthz(w http.ResponseWriter, _ *http.Request) {
-	w.WriteHeader(http.StatusOK)
-}
-
-// handleIndex renders the full page shell with root-level tree nodes.
-func (s *Server) handleIndex(w http.ResponseWriter, r *http.Request) {
-	entities, err := s.cfg.App.ListEntities(r.Context())
-	if err != nil {
-		s.serverError(w, "list entities", err)
-		return
-	}
-
-	var roots []app.EntityResult
-	for _, e := range entities {
-		if isRootEntity(e) {
-			roots = append(roots, e)
-		}
-	}
-
-	data := indexData{
-		Roots:     roots,
-		Version:   version.Version,
-		GitCommit: version.GitCommit,
-	}
-	s.renderHTML(w, "index", data)
-}
-
-// handleTreeChildren returns the direct children of an entity as tree_node fragments.
-func (s *Server) handleTreeChildren(w http.ResponseWriter, r *http.Request) {
-	entityID := r.PathValue("entityID")
-
-	if _, err := s.cfg.App.GetEntityByID(r.Context(), entityID); err != nil {
-		if errors.Is(err, store.ErrNotFound) {
-			http.Error(w, "entity not found", http.StatusNotFound)
-			return
-		}
-		s.serverError(w, "get entity", err)
-		return
-	}
-
-	children, err := s.cfg.App.GetChildren(r.Context(), entityID)
-	if err != nil {
-		s.serverError(w, "get children", err)
-		return
-	}
-
-	var buf bytes.Buffer
-	for _, child := range children {
-		if tmplErr := s.templates.ExecuteTemplate(&buf, "tree_node", child); tmplErr != nil {
-			s.serverError(w, "execute tree_node", tmplErr)
-			return
-		}
-	}
-	w.Header().Set("Content-Type", "text/html; charset=utf-8")
-	if _, werr := w.Write(buf.Bytes()); werr != nil {
-		s.cfg.Logger.Error("write tree children", "error", werr)
-	}
-}
-
-type indexData struct {
-	Roots     []app.EntityResult
-	Version   string
-	GitCommit string
-}
-
-// handleEntityDetail returns the detail pane fragment for a single entity.
+// detailData is the view model for the entity detail pane.
 type detailData struct {
 	Entity         app.EntityResult
 	DateAdded      string
@@ -330,142 +217,6 @@ func (s *Server) handleEditName(w http.ResponseWriter, r *http.Request) {
 	if r.Header.Get("Hx-Request") == htmxHeaderVal {
 		s.renderHTMLAppend(w, "tree_label_oob", fresh.Entity)
 	}
-}
-
-// isRootEntity returns true when e has no parent (no colon in path = depth 0).
-func isRootEntity(e app.EntityResult) bool {
-	return !strings.Contains(e.FullPathDisplay, ":")
-}
-
-// handleAddItemForm returns an inline modal form for adding a child entity under parentID.
-func (s *Server) handleAddItemForm(w http.ResponseWriter, r *http.Request) {
-	s.renderAddForm(w, r, r.PathValue("parentID"))
-}
-
-// handleRootAddItemForm serves the add-item form for creating a root-level entity.
-func (s *Server) handleRootAddItemForm(w http.ResponseWriter, r *http.Request) {
-	s.renderAddForm(w, r, "")
-}
-
-type addFormData struct {
-	ParentID string // empty means root
-	Target   string // CSS selector for hx-target
-}
-
-func (s *Server) renderAddForm(w http.ResponseWriter, _ *http.Request, parentID string) {
-	target := "#tree-children-" + parentID
-	if parentID == "" {
-		target = "#tree-root-list"
-	}
-	s.renderHTML(w, "add_item_form", addFormData{ParentID: parentID, Target: target})
-}
-
-// handleAddItem processes the POST form, creates the entity, and returns a fresh tree_node fragment.
-func (s *Server) handleAddItem(w http.ResponseWriter, r *http.Request) {
-	parentID := r.PathValue("parentID")
-
-	parent, err := s.cfg.App.GetEntityByID(r.Context(), parentID)
-	if err != nil {
-		if errors.Is(err, store.ErrNotFound) {
-			http.Error(w, "parent entity not found", http.StatusNotFound)
-			return
-		}
-		s.serverError(w, "get parent entity", err)
-		return
-	}
-
-	s.createAndRenderNode(w, r, parent.FullPathDisplay)
-}
-
-// handleRootAddItem processes the POST form for a new root-level entity.
-func (s *Server) handleRootAddItem(w http.ResponseWriter, r *http.Request) {
-	s.createAndRenderNode(w, r, "")
-}
-
-func (s *Server) createAndRenderNode(w http.ResponseWriter, r *http.Request, parentPath string) {
-	if err := r.ParseForm(); err != nil {
-		http.Error(w, "bad form", http.StatusBadRequest)
-		return
-	}
-
-	displayName := strings.TrimSpace(r.FormValue("display_name"))
-	if displayName == "" {
-		http.Error(w, "display_name is required", http.StatusBadRequest)
-		return
-	}
-
-	entityType, err := inventory.ParseEntityType(r.FormValue("entity_type"))
-	if err != nil {
-		http.Error(w, "invalid entity_type", http.StatusBadRequest)
-		return
-	}
-
-	actor := strings.TrimSpace(r.FormValue("user"))
-	if actor == "" {
-		actor = "webui"
-	}
-
-	created, err := s.cfg.App.CreateEntity(r.Context(), app.CreateEntityRequest{
-		DisplayName: displayName,
-		EntityType:  entityType,
-		ParentPath:  parentPath,
-		ActorID:     actor,
-	})
-	if err != nil {
-		s.serverError(w, "create entity", err)
-		return
-	}
-
-	s.renderHTML(w, "tree_node", created)
-}
-
-// handleSearch handles GET /search?q=... returning either a full page or
-// an HTMX fragment depending on the Hx-Request header.
-func (s *Server) handleSearch(w http.ResponseWriter, r *http.Request) {
-	q := strings.TrimSpace(r.URL.Query().Get("q"))
-	if len(q) > maxQueryLen {
-		http.Error(w, "query too long", http.StatusBadRequest)
-		return
-	}
-
-	htmx := r.Header.Get("Hx-Request") == htmxHeaderVal
-
-	var searchData any
-	if q != "" {
-		results, err := s.cfg.App.FindEntities(r.Context(), app.FindEntitiesRequest{
-			Query: q,
-			Limit: searchLimit,
-		})
-		if err != nil {
-			s.serverError(w, "search", err)
-			return
-		}
-		searchData = struct {
-			Query   string
-			Results []app.FindResult
-		}{Query: q, Results: results}
-	}
-
-	if htmx {
-		s.renderHTML(w, "search_results", searchData)
-		return
-	}
-
-	entities, err := s.cfg.App.ListEntities(r.Context())
-	if err != nil {
-		s.serverError(w, "list entities", err)
-		return
-	}
-	var roots []app.EntityResult
-	for _, e := range entities {
-		if isRootEntity(e) {
-			roots = append(roots, e)
-		}
-	}
-	s.renderHTML(w, "search_page", struct {
-		Roots  []app.EntityResult
-		Search any
-	}{Roots: roots, Search: searchData})
 }
 
 // handleToggleMissing handles POST /entities/{entityID}/actions/toggle-missing.
